+++ src/slskd/VirtualSoulfind/v2/Catalogue/TrackCopyState.cs
@@ -1,0 +1,6 @@
+// <copyright file="TrackCopyState.cs" company="slskdN Team">
+//     Copyright (c) slskdN Team. All rights reserved.
+// </copyright>
+namespace slskd.VirtualSoulfind.v2.Catalogue;
+
+public readonly record struct TrackCopyState(bool HasLocalFile, bool HasVerifiedCopy);